Allow date strings for #inst to have milliseconds
JSON dates usually are encoded like: "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'" not
"yyyy-MM-dd'T'HH:mm:ss'Z'". For backwards compatibility allow both.

Needed to add midje as a dependency so tests run.

Needed to update to clojure 1.8.0 for the new datomic. The error was:
java.lang.NoClassDefFoundError: clojure/lang/Tuple

Needed to revert back to version 2.2.17 of the hara libraries because of
java.lang.IllegalAccessError with version 2.5.2.